highlighter/requirements.txt
ipykernel
pandas
lightgbm
numpy
python-chess